Always clear the in_progress task
This is the main cause of the bug fixed in the previous commit - it just
only ocurred to me in bed last night. We throw an exception when waiting
on the task, which means we never actually clear the in_progress field.
Thus once a task throws an exception once, its stuck in that state
forever.